+++ packages/terra-framework-docs/package.json
@@ -47,6 +47,7 @@
     "terra-form-input": "^4.4.0",
     "terra-form-select": "^6.8.0",
     "terra-grid": "^6.0.0",
+    "terra-menu": "^6.60.0",
     "terra-modal-manager": "^6.58.0",
     "terra-section-header": "^2.0.0",
     "terra-slide-panel-manager": "^5.61.0"
