+++ packages/terra-framework-docs/src/terra-dev-site/doc/menu/Menu.1.doc.mdx
@@ -4,11 +4,11 @@ import MenuPropsTable from 'terra-menu/src/Menu?dev-site-props-table';
 import MenuItemPropsTable from 'terra-menu/src/MenuItem?dev-site-props-table';
 import MenuItemGroupPropsTable from 'terra-menu/src/MenuItemGroup?dev-site-props-table';
 
-import ArrowMenu from '../example/ArrowMenu?dev-site-example';
-import BasicMenu from '../example/BasicMenu?dev-site-example';
-import MenuBounded from '../example/MenuBounded?dev-site-example';
-import MenuWidths from '../example/MenuWidths?dev-site-example';
-import MenuWithInstructionsForUse from '../example/MenuWithInstructionsForUse?dev-site-example';
+import ArrowMenu from './example/ArrowMenu?dev-site-example';
+import BasicMenu from './example/BasicMenu?dev-site-example';
+import MenuBounded from './example/MenuBounded?dev-site-example';
+import MenuWidths from './example/MenuWidths?dev-site-example';
+import MenuWithInstructionsForUse from './example/MenuWithInstructionsForUse?dev-site-example';
 
 <Badge />
