+++ packages/terra-framework-docs/src/terra-dev-site/test/menu/menu/BoundedMenu.test.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import classNames from 'classnames/bind';
-import Menu from '../../../../Menu';
+import Menu from 'terra-menu';
 import styles from './BoundedMenu.test.module.scss';
 
 const cx = classNames.bind(styles);
