+++ packages/terra-framework-docs/src/terra-dev-site/test/menu/menu/DefaultMenu.test.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
-import Menu from '../../../../Menu';
+import Menu from 'terra-menu';
 
-class SelectableMenu extends React.Component {
+class DefaultMenu extends React.Component {
   constructor(props) {
     super(props);
     this.handleButtonClick = this.handleButtonClick.bind(this);
@@ -9,6 +9,10 @@ class SelectableMenu extends React.Component {
     this.setButtonNode = this.setButtonNode.bind(this);
     this.getButtonNode = this.getButtonNode.bind(this);
     this.state = { open: false };
+  }
+
+  componentDidMount() {
+    this.forceUpdate();
   }
 
   handleButtonClick() {
@@ -30,21 +34,14 @@ class SelectableMenu extends React.Component {
   render() {
     return (
       <div>
-        <div>
-          This menu contains one selectable child (an Item Group). The menu should allow space for the checkmark and the items in the item group should be selectable.
-        </div>
         <Menu
           isOpen={this.state.open}
           targetRef={this.getButtonNode}
           onRequestClose={this.handleRequestClose}
         >
-          <Menu.ItemGroup className="TestGroup" onChange={this.handleSelection} key="selectable-group">
-            <Menu.Item text="Group Item 1" key="1" className="TestGroupItem1" />
-            <Menu.Item text="Group Item 2" key="2" className="TestGroupItem2" />
-            <Menu.Item text="Group Item 3" key="3" className="TestGroupItem3" />
-          </Menu.ItemGroup>
+          <Menu.Item text="Default Menu" id="TestContent" />
         </Menu>
-        <button type="button" id="selectable-menu-button" onClick={this.handleButtonClick} ref={this.setButtonNode}>
+        <button type="button" id="default-button" onClick={this.handleButtonClick} ref={this.setButtonNode}>
           Default Menu
         </button>
       </div>
@@ -52,4 +49,4 @@ class SelectableMenu extends React.Component {
   }
 }
 
-export default SelectableMenu;
+export default DefaultMenu;
