+++ packages/terra-framework-docs/src/terra-dev-site/test/menu/menu/DynamicMenuChildren.test.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Menu from '../../../../Menu';
+import Menu from 'terra-menu';
 
 class DefaultMenu extends React.Component {
   constructor(props) {
